Create runTreeReader.C
Macro to run over multiple samples and generate the plots associated with the samples. You need an "outputroot" directory to run the macro and the input samples must ahve the structure "Ntuple" => 50_test is the sample that you want. You can of course change that. Basically creates a root file and a txt file that summarizes the different efficiencies of the selections